requirements.txt
click>=8.1.0
google-generativeai>=0.8.0
manim>=0.18.0
loguru>=0.7.0
python-dotenv>=1.0.0